barcode/requirements.txt
androidstorage4kivy==0.1.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
idna==3.10
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
plyer==2.1.0
Pygments==2.18.0
python-barcode==0.15.1
qrcode==8.0
requests==2.32.3
segno==1.6.1
soupsieve==2.6
urllib3==2.2.3
watchdog==6.0.0
